+++ src/tables/mod.rs
@@ -30,4 +30,6 @@ pub mod os2;
 pub mod post;
 pub mod sbix;
 pub mod stat;
+pub mod vhea;
+pub mod vmtx;
 pub mod vvar;